debian: add Breaks/Conflicts/Replaces to overwrite old version of `camera-streamer` (#79)
As discussed on Discord a while ago, adding
Breaks/Conflicts/Replaces entries against
camera-streamer (<< 0.2) allows for a smooth
migration from the camera-streamer package
built for OctoPi to camera-streamer-raspi, through
a transitional camera-streamer version 0.2 on
OctoPi's side.

An upgrade of camera-streamer from OctoPrint's
apt repo will pull in an updated transitional
camera-streamer package version 0.2 that depends
on camera-streamer-raspi. For that to install
flawlessly in all cases, it needs to be installed
*before* the camera-streamer-raspi dependency
(as some files between that and the former
camera-streamer package overlap). By
declaring Breaks/Conflicts/Replaces relationships
on camera-streamer-raspi this is ensured.