build: Support building new revisions for tags
A new action has been created in build_revision.yaml with a
workflow dispatch that takes two optional input parameters,
tag and revision.

If provided, the release with the tag will be adjusted with
new builds versionened with the revision.

Otherwise the latest tag is determined, its release information
fetched and the current package revision determined from that,
then incremented by one to act as the new revision.

The goal of this action is to be able to do additional
builds of the *same* code version, against a newer dependency
versions (e.g. on a libcamera0 update).